feat: add clap
feat: working
