+++ Cargo.toml
@@ -30,7 +30,7 @@ actix-service = "2.0.0"
 actix-web = "4.2"
 async-trait = "0.1"
 chrono = "0.4.23"
-cron = "0.12.0"
+clap = { version = "4.1.4", features = ["derive"] }
 futures-util = "0.3.25"
 lazy_static = "1.4.0"
 log = "0.4.17"
